Rename generateId to GenerateId and write tests for it
