+++ packages/components/src/inputs/TextArea.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import FieldGroup, { FieldGroupDefaultProps, FieldGroupProps } from "../field-group/FieldGroup";
 // Import the required CSS.
 import "./input.css";
-import { getId } from "../generateId";
+import { getId } from "../GenerateId";
 
 /**
  * Renders a textarea for use in our HTML forms.
